bldc functions, pin bug

+++ BLDC/Commutation.h
@@ -32,8 +32,8 @@ typedef enum
 
 typedef struct 
 {
-	void (*SetPhasePWM)(uint8_t pwm);
-	void (*ActivatePhase)(uint8_t pwm);
+	void (*SetPhasePWM)(uint16_t pwm);
+	void (*ActivatePhase)(uint16_t pwm);
 	//PhaseID?
 } BLDC_COMMUTATION_PHASE_T;
 
@@ -92,11 +92,11 @@ typedef struct
 BLDC_COMMUTATION_T;
 
 
-extern void BLDC_Commutation_ISR(BLDC_COMMUTATION_T * commutation, uint8_t pwm);
+extern void BLDC_Commutation_ISR(BLDC_COMMUTATION_T * commutation, uint16_t pwm);
 
-bool BLDC_Commutation_Poll(BLDC_COMMUTATION_T * commutation, uint8_t pwm);
-void (*BLDC_Commutation_GetFunctionActivatePhase(BLDC_COMMUTATION_T * commutation))(uint8_t);
-void (*BLDC_Commutation_GetFunctionSetPhasePWM(BLDC_COMMUTATION_T * commutation))(uint8_t);
+bool BLDC_Commutation_Poll(BLDC_COMMUTATION_T * commutation, uint16_t pwm);
+void (*BLDC_Commutation_GetFunctionActivatePhase(BLDC_COMMUTATION_T * commutation))(uint16_t);
+void (*BLDC_Commutation_GetFunctionSetPhasePWM(BLDC_COMMUTATION_T * commutation))(uint16_t);
 void BLDC_Commutation_SetDirection(BLDC_COMMUTATION_T * commutation, BLDC_DIRECTION_T dir);
 
 void BLDC_Commutation_MapCommuntationTableRunCalibration
@@ -108,19 +108,19 @@ void BLDC_Commutation_MapCommuntationTableRunCalibration
 	uint8_t * returnIndexBA,
 	uint8_t * returnIndexCA,
 	uint8_t * returnIndexCB,
-	void (*setPWMPhaseAB)(uint8_t),
-	void (*setPWMPhaseAC)(uint8_t),
-	void (*setPWMPhaseBC)(uint8_t),
-	void (*setPWMPhaseBA)(uint8_t),
-	void (*setPWMPhaseCA)(uint8_t),
-	void (*setPWMPhaseCB)(uint8_t),
+	void (*setPWMPhaseAB)(uint16_t),
+	void (*setPWMPhaseAC)(uint16_t),
+	void (*setPWMPhaseBC)(uint16_t),
+	void (*setPWMPhaseBA)(uint16_t),
+	void (*setPWMPhaseCA)(uint16_t),
+	void (*setPWMPhaseCB)(uint16_t),
 	uint16_t pwm,
-	void (*commutatePhaseAB)(uint8_t),
-	void (*commutatePhaseAC)(uint8_t),
-	void (*commutatePhaseBC)(uint8_t),
-	void (*commutatePhaseBA)(uint8_t),
-	void (*commutatePhaseCA)(uint8_t),
-	void (*commutatePhaseCB)(uint8_t),
+	void (*commutatePhaseAB)(uint16_t),
+	void (*commutatePhaseAC)(uint16_t),
+	void (*commutatePhaseBC)(uint16_t),
+	void (*commutatePhaseBA)(uint16_t),
+	void (*commutatePhaseCA)(uint16_t),
+	void (*commutatePhaseCB)(uint16_t),
 	void (*enablePhaseABC)(void),
 	void (*delay)(uint32_t),
 	uint32_t delayTime
@@ -129,18 +129,18 @@ void BLDC_Commutation_MapCommuntationTableRunCalibration
 void BLDC_Commutation_MapCommuntationTable
 (	
 	BLDC_COMMUTATION_T * commutation,				
-	void (*setPWMPhaseAB)(uint8_t),
-	void (*setPWMPhaseAC)(uint8_t),
-	void (*setPWMPhaseBC)(uint8_t),
-	void (*setPWMPhaseBA)(uint8_t),
-	void (*setPWMPhaseCA)(uint8_t),
-	void (*setPWMPhaseCB)(uint8_t),
-	void (*commutatePhaseAB)(uint8_t),
-	void (*commutatePhaseAC)(uint8_t),
-	void (*commutatePhaseBC)(uint8_t),
-	void (*commutatePhaseBA)(uint8_t),
-	void (*commutatePhaseCA)(uint8_t),
-	void (*commutatePhaseCB)(uint8_t),
+	void (*setPWMPhaseAB)(uint16_t),
+	void (*setPWMPhaseAC)(uint16_t),
+	void (*setPWMPhaseBC)(uint16_t),
+	void (*setPWMPhaseBA)(uint16_t),
+	void (*setPWMPhaseCA)(uint16_t),
+	void (*setPWMPhaseCB)(uint16_t),
+	void (*commutatePhaseAB)(uint16_t),
+	void (*commutatePhaseAC)(uint16_t),
+	void (*commutatePhaseBC)(uint16_t),
+	void (*commutatePhaseBA)(uint16_t),
+	void (*commutatePhaseCA)(uint16_t),
+	void (*commutatePhaseCB)(uint16_t),
 	uint8_t indexPhaseAB,
 	uint8_t indexPhaseAC,
 	uint8_t indexPhaseBC,
@@ -154,18 +154,18 @@ void BLDC_Commutation_Init
 	BLDC_COMMUTATION_T * commutation,
 	BLDC_DIRECTION_T dir,
 	uint8_t (*getHallSensors)(void),
-	void (*setPWMPhaseAB)(uint8_t),
-	void (*setPWMPhaseAC)(uint8_t),
-	void (*setPWMPhaseBC)(uint8_t),
-	void (*setPWMPhaseBA)(uint8_t),
-	void (*setPWMPhaseCA)(uint8_t),
-	void (*setPWMPhaseCB)(uint8_t),
-	void (*commutatePhaseAB)(uint8_t),
-	void (*commutatePhaseAC)(uint8_t),
-	void (*commutatePhaseBC)(uint8_t),
-	void (*commutatePhaseBA)(uint8_t),
-	void (*commutatePhaseCA)(uint8_t),
-	void (*commutatePhaseCB)(uint8_t),
+	void (*setPWMPhaseAB)(uint16_t),
+	void (*setPWMPhaseAC)(uint16_t),
+	void (*setPWMPhaseBC)(uint16_t),
+	void (*setPWMPhaseBA)(uint16_t),
+	void (*setPWMPhaseCA)(uint16_t),
+	void (*setPWMPhaseCB)(uint16_t),
+	void (*commutatePhaseAB)(uint16_t),
+	void (*commutatePhaseAC)(uint16_t),
+	void (*commutatePhaseBC)(uint16_t),
+	void (*commutatePhaseBA)(uint16_t),
+	void (*commutatePhaseCA)(uint16_t),
+	void (*commutatePhaseCB)(uint16_t),
 	uint8_t indexPhaseAB,
 	uint8_t indexPhaseAC,
 	uint8_t indexPhaseBC,
